refactor: Unify spanConfig save function
