+++ src/lib/component/SettingDialog/reflectImmediately/saveSpanConfig/unique.js
